pyproject.toml
[project]
name = "simulate-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.7.2",
    "dataclasses>=0.8",
    "openai-agents>=0.3.0",
    "py-dotenv>=0.1",
    "pydantic>=2.11.7",
    "streamlit>=1.49.1",
    "typing>=3.10.0.0",
]
